fix: verify badly configured reverse proxy + more complete error messages
Now throws EMISSINGREPO only when response with statusCode at 404 and statusMessage is "Project Not
Found" and throws EUNEXPECTED404 when there is a problem with a badly configured reverse-proxy
(similar response but with statusMessage is "Not Found"). Also adds information when error is thrown
to help with troubleshooting.

fix #99 and re #92